major refactor of library structure
Abstract routes, middleware & API controllers into separate files
