+++ lib/utils.js
@@ -3,7 +3,7 @@ const moment = require("moment");
 const crypto = require("crypto");
 const geoip = require("geoip-lite");
 
-const database = require("./db");
+const database = require("./db/db");
 
 /**
  * Creates required collections for bugcatch to operate
